requirements.txt
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
deprecation==2.1.0
firebase-admin==4.3.0
Flask==1.0.2
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.7.0
google-cloud-secret-manager==1.0.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.30.0
httplib2==0.18.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpickle==0.9.6
MarkupSafe==1.1.1
msgpack==1.0.0
packaging==20.4
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
squareup==6.0.0.20200625
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
